Troubleshoot weasel repl.
Weasel uses the clojurescript compiler to analyze source for completion
or something.  Analysis seems to depend on all clojurescript being under
a single directory and for cljs directories to be listed in the 
top level :source-paths key of the leingingen project.clj.

Also renamed resolve to resolve-url to keep it from conflicting with 
clojure.core/resolve.

https://github.com/tomjakubowski/weasel/issues/28